% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_acml.R
\name{ACi1q}
\alias{ACi1q}
\title{Ascertainment correction piece for univariate sampling}
\usage{
ACi1q(cutpoints, SampProb, mu_q, sigma_q)
}
\arguments{
\item{cutpoints}{cutpoints defining the sampling regions. (a vector of length 2)}

\item{SampProb}{Sampling probabilities from within each region (vector of length 3).}

\item{mu_q}{a scalar for the mean value of the Q_i distribution}

\item{sigma_q}{a scalar for the standard deviation of the Q_i distribution}
}
\value{
Not yet log transformed ascertainment correction
}
\description{
Calculate the (not yet log transformed) ascertainment correction under a univariate Q_i
}
