% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{ACi2q}
\alias{ACi2q}
\title{Ascertainment correction piece for bivariate sampling}
\usage{
ACi2q(cutpoints, SampProb, mu_q, sigma_q)
}
\arguments{
\item{cutpoints}{cutpoints defining the sampling regions. (a vector of length 4: c(xlow, xhigh, ylow, yhigh))}

\item{SampProb}{Sampling probabilities from within each of two sampling regions; central region and outlying region (vector of length 2).}

\item{mu_q}{a 2-vector for the mean value of the bivariate Q_i distribution.}

\item{sigma_q}{a 2 by 2 covariance matrix for the bivariate Q_i distribution.}
}
\value{
Not yet log transformed ascertainment correction
}
\description{
Calculate the (not yet log transformed) ascertainment correction under a bivariate Q_i
}
