% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{fixef}
\alias{fixef}
\alias{ranef}
\alias{coef.acml}
\title{Extract parameters from fitted models}
\usage{
fixef(object, ...)

ranef(object, transform = FALSE, ...)

\method{coef}{acml}(object, complete = TRUE, transform = FALSE, ...)
}
\arguments{
\item{object}{the fitted model object to extract model coefficients.}

\item{...}{Additional arguments passed along}

\item{transform}{logical(1); If TRUE the coefficients will be inverse
transformed back to their original scale, this is `exp` for deviation
components and Fisher transformed for correlation.}

\item{complete}{Not used at present, required for S3}
}
\value{
A named vector of desired coeffients.
}
\description{
These are S3 methods to extract the entire parameter set, just the fixed
effects, or just the random effects. They are by default returned on the
unconstrained optimization scale.
}
\examples{
data(gbti)
design <- ods(Response ~ Month|Patient, 'intercept', p_sample=c(1, 0.25, 1),
              data=gbti, quantiles=c(0.1, 0.9))
est <- acml(Response ~ Month*Genotype, design, gbti)
coef(est)
ranef(est)
fixef(est)
}
