% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\name{CreateSubjectDataWL}
\alias{CreateSubjectDataWL}
\title{Create a list of subject-specific data}
\usage{
CreateSubjectDataWL(id, y, x, z, Weights)
}
\arguments{
\item{id}{sum(n_i) vector of subject ids}

\item{y}{response vector}

\item{x}{sum(n_i) by p design matrix for fixed effects}

\item{z}{sum(n_i) by 2 design matric for random effects (intercept and slope)}

\item{Weights}{Subject specific sampling weights.  A vector of length sum(n_i).  Not used unless using weighted Likelihood}
}
\description{
Create a list of subject-specific data
}
