% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{li.lme}
\alias{li.lme}
\title{Calculate a subject-specific contribution to a log-likelihood for longitudinal normal data}
\usage{
li.lme(yi, xi, beta, vi)
}
\arguments{
\item{yi}{n_i-response vector}

\item{xi}{n_i by p design matrix for fixed effects}

\item{beta}{mean model parameter vector}

\item{vi}{the variance covariance matrix (ZDZ+Sige2*I)}
}
\value{
subject specific contribution to the log-likelihood
}
\description{
Calculate a subject-specific contribution to a log-likelihood for longitudinal normal data
}
