% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{li.lme.score2}
\alias{li.lme.score2}
\title{Subject specific contribution to the lme model score (also returns marginal Vi=Cov(Y|X))}
\usage{
li.lme.score2(subjectData, beta, sigma.vc, rho.vc, sigma.e)
}
\arguments{
\item{subjectData}{a list that contains yi, xi, zi, Weights.i.  Note that Weights.i is used for inverse probability weighting only.}

\item{beta}{mean model parameter p-vector}

\item{sigma.vc}{vector of variance components on standard deviation scale}

\item{rho.vc}{vector of correlations among the random effects.  The length should be q choose 2}

\item{sigma.e}{std dev of the measurement error distribution}
}
\value{
Subject specific contribution to the log-likelihood score (also returns marginal Vi=Cov(Y|X))
}
\description{
Subject specific contribution to the lme model score (also returns marginal Vi=Cov(Y|X))
}
