% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions5.R, R/acml.R
\name{logACi2q}
\alias{logACi2q}
\title{Log of the Ascertainment correction piece for bivariate sampling}
\usage{
logACi2q(
  yi,
  xi,
  zi,
  wi,
  beta,
  sigma.vc,
  rho.vc,
  sigma.e,
  cutpoints,
  SampProb,
  mui.phase1
)

logACi2q(
  yi,
  xi,
  zi,
  wi,
  beta,
  sigma.vc,
  rho.vc,
  sigma.e,
  cutpoints,
  SampProb,
  mui.phase1
)
}
\arguments{
\item{yi}{n_i-response vector}

\item{xi}{n_i by p design matrix for fixed effects}

\item{zi}{n_i by q design matric for random effects (intercept and slope)}

\item{wi}{the pre-multiplier of yi to generate the sampling variable q_i}

\item{beta}{mean model parameter p-vector}

\item{sigma.vc}{vector of variance components on standard deviation scale}

\item{rho.vc}{vector of correlations among the random effects.  The length should be q choose 2}

\item{sigma.e}{std dev of the measurement error distribution}

\item{cutpoints}{cutpoints defining the sampling regions. (a vector of length 4 c(xlow, xhigh, ylow, yhigh))}

\item{SampProb}{Sampling probabilities from within each region (vector of length 2 c(central region, outlying region)).}

\item{mui.phase1}{linear predictor from phase 1 that is used to calculate mu_q if we are doing BLUP based sampling.  If doing ODS then this is a vector of 0s}
}
\value{
log transformed ascertainment correction

log transformed ascertainment correction
}
\description{
Calculate the log transformed ascertainment correction under a bivariate Q_i.  Also return vi

Calculate the log transformed ascertainment correction under a bivariate Q_i.  Also return vi
}
