% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\name{LogLikeC.ScoreWL}
\alias{LogLikeC.ScoreWL}
\title{Calculate the gradient of the conditional likelihood}
\usage{
LogLikeC.ScoreWL(
  beta,
  sigma.vc,
  rho.vc,
  sigma.e,
  CheeseCalc = FALSE,
  subjectData
)
}
\arguments{
\item{beta}{mean model parameter p-vector}

\item{sigma.vc}{vector of variance components on standard deviation scale}

\item{rho.vc}{vector of correlations among the random effects.  The length should be q choose 2}

\item{sigma.e}{std dev of the measurement error distribution}

\item{CheeseCalc}{If FALSE, the function returns the gradient of the conditional log likelihood across all subjects.  If TRUE, the cheese part of the sandwich esitmator is calculated.}

\item{subjectData}{Precomputed subject-specific data list from \code{CreateSubjectDataWL}.}
}
\value{
If CheeseCalc=FALSE, gradient of conditional log likelihood.  If CheeseCalc=TRUE, the cheese part of the sandwich estimator is calculated.
}
\description{
Calculate the gradient of the conditional likelihood for the univariate and bivariate sampling cases across all subjects (CheeseCalc=FALSE) or the cheese part of the sandwich estimator if CheeseCalc=TRUE.
}
