% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\name{LogLikeiWL}
\alias{LogLikeiWL}
\title{Calculate the ss contributions to the conditional likelihood for the univariate and bivariate sampling cases.}
\usage{
LogLikeiWL(subjectData, beta, sigma.vc, rho.vc, sigma.e)
}
\arguments{
\item{subjectData}{a list containing: yi, xi, zi, Weights.i}

\item{beta}{mean model parameter p-vector}

\item{sigma.vc}{vector of variance components on standard deviation scale}

\item{rho.vc}{vector of correlations among the random effects.  The length should be q choose 2}

\item{sigma.e}{std dev of the measurement error distribution}
}
\value{
ss contributions to the conditional log likelihood.  This is an internal function used by LogLikeWL
}
\description{
Calculate the ss contributions to the conditional likelihood for the univariate and bivariate sampling cases.
}
