% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods.R
\name{ods}
\alias{ods}
\title{Specify a given design for Outcome Dependent Sampling (ODS)}
\usage{
ods(
  formula,
  method = NULL,
  p_sample = NULL,
  weights = NULL,
  data = NULL,
  quantiles = NULL,
  cutpoints = NULL,
  subset = NULL,
  prob_intercept = NULL,
  method_name = NULL,
  acml_samp_prob_name = NULL,
  cutpoints_name = NULL,
  sampled_name = NULL,
  na.action = getOption("na.action"),
  ProfileCol = NULL,
  ...
)
}
\arguments{
\item{formula}{(or an object that can be
coerced to that class): a symbolic description of the sampling
model.}

\item{method}{`character(1)`; A string that specifies 'slope', 'intercept', 'bivariate', or 'mean'.}

\item{p_sample}{`numeric(n)`; A numeric vector with sampling probabilities
for each region specified by cutpoints.}

\item{weights}{The column name of the sampling weights for each subject, which is the inverse of the sampling probability of the subject being selected.}

\item{data}{`data.frame`; an optional data frame, list or environment (or
object coercible by as.data.frame to a data frame) containing the variables
in the model. If not found in data, the variables are taken from
environment(formula), typically the environment from which acml is called.}

\item{quantiles}{`numeric(n-1)`; quantiles to compute cutpoints of interest.
One of quantiles or cutpoints must be specified. Default is NULL.}

\item{cutpoints}{`numeric(n-1)`; a specification of the cutpoints for sampling. Default is NULL.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process. (See additional details about how this
argument interacts with data-dependent bases in the ‘Details’ section of
the model.frame documentation.)}

\item{prob_intercept}{The proportion of selecting the Phase-II sample in the mixture design of two-phase sampling.}

\item{method_name}{The column name of the method. This is ONLY used when the two-phase sampling was completed and the Phase-II sample was collected. When using this field, the p_sample, quantiles, and cutpoints field should not be used.}

\item{acml_samp_prob_name}{The column name of the sampling probabilities from within each region. For low medium high sampling, each cell is a vector of length 3 with sampling probabilities for each region. For bivariate stratum sampling each cell is a vector of length 2 with sampling probabilities for the inner and outer strata. Each subject should have n_i rows of the same values. This is ONLY used when the two-phase sampling was completed and the Phase-II sample was collected. When using this field, the p_sample, quantiles, and cutpoints field should not be used.}

\item{cutpoints_name}{The column name of the cutpoints for each region. This is ONLY used when the two-phase sampling was completed and the Phase-II sample was collected. When using this field, the p_sample, quantiles, and cutpoints field should not be used.}

\item{sampled_name}{The column name of whether the subject is sampled in Phase-II data collection. This is ONLY used when the two-phase sampling was completed and the Phase-II sample was collected. When using this field, the p_sample, quantiles, and cutpoints field should not be used.}

\item{na.action}{a function which indicates what should happen when the data
contain NAs. The default is set by the na.action setting of options, and is
na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another
possible value is NULL, no action. Value na.exclude can be useful.}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maimizing the log likelihood for all other parameters while fixing these columns at the values of params at the location of ProfileCol}

\item{...}{additional arguments.}
}
\value{
Returns an ODS design object.
}
\description{
Specify the design of an outcome dependent sampling routine.
}
\examples{
data(gbti)

odsd <- ods(Response ~ Month|Patient, 'mean', p_sample=c(1, 0.25, 1),
            data=gbti, quantiles=c(0.1, 0.9))
summary(odsd)
plot(odsd)

odsd <- ods(Response ~ Month|Patient, 'intercept', p_sample=c(1, 0.25, 1),
            data=gbti, quantiles=c(0.1, 0.9))
summary(odsd)
plot(odsd)

odsd <- ods(Response ~ Month|Patient, 'slope', p_sample=c(1, 0.25, 1),
            data=gbti, quantiles=c(0.1, 0.9))
summary(odsd)
plot(odsd)

odsd <- ods(Response ~ Month|Patient, 'bivariate', p_sample=c(0.25, 1),
            data=gbti, quantiles=0.8)
summary(odsd)
plot(odsd)

}
\seealso{
[plot.odsdesign()]
}
