% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{robcov}
\alias{robcov}
\title{robcov: Return the robust covariance matrix.}
\usage{
robcov(object, ...)
}
\arguments{
\item{object}{to get robust variance covariance matrix.}

\item{...}{other arguments}
}
\description{
Retrieve the robust variance covariance matrix
}
\details{
An S3 Method to retrieve the robust variance-covariance matrix of an ODS
model. It is derived via the sandwich estimator.
}
