% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\name{WL}
\alias{WL}
\title{Fit model using ascertainment corrected likelihood model (ACML)}
\usage{
WL(
  formula,
  design,
  subset = NULL,
  na.action = getOption("na.action"),
  verbose = 0L,
  init = NULL,
  ProfileCol = NA,
  ...
)
}
\arguments{
\item{formula}{`formula`; (or an object that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of model specification are given under ‘Details’.}

\item{design}{an object of class 'odsdesign'. This specifies the
design of sampling used for the fitting algorithm.}

\item{subset}{`logical`; an optional vector specifying a subset of
observations to be used in the fitting process. (See additional details
about how this argument interacts with data-dependent bases in the
‘Details’ section of the model.frame documentation.)}

\item{na.action}{`function`; an optional vector specifying a subset of
observations to be used in the fitting process. (See additional details
about how this argument interacts with data-dependent bases in the
‘Details’ section of the model.frame documentation.)}

\item{verbose}{`numeric(1)`; Debugging information printing level
(sent to the optimizer).}

\item{init}{`numeric`; Initial starting position for parameter search
via the optimizer.}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maimizing the log likelihood for all other parameters while fixing these columns at the values of InitVals[ProfileCol]}

\item{...}{Optional additional parameters passed to sub methods.}
}
\description{
Outcome dependent sampling designs need to be corrected when fitting a
statistical model for proper inferences. This routine will fit and return
the model fit using ascertainment corrected likelihood.
}
\examples{
data(gbti)
design <- ods(Response ~ Month|Patient, 'intercept', p_sample=c(1, 0.25, 1),
              data=gbti, quantiles=c(0.1, 0.9), weights="pSample")
est <- WL(Response ~ Month, design)
est
summary(est)
}
