% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\name{WL_internal}
\alias{WL_internal}
\title{Fitting function: ACML or WL for a linear mixed effects model (random intercept and slope)}
\usage{
WL_internal(formula, design, InitVals, ProfileCol)
}
\arguments{
\item{formula}{is the outcome model formula used for the linear mixed model.}

\item{design}{an object of class 'odsdesign'. This specifies the design of sampling used for the fitting algorithm.}

\item{InitVals}{starting values for c(beta, log(sigma0), log(sigma1), log((1+rho)/(1-rho)), log(sigmae))}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maimizing the log likelihood for all other parameters while fixing these columns at the values of InitVals[ProfileCol]}
}
\value{
Weighted likelihood: Ests, covar, logLik, code, robcov
}
\description{
Fitting function: ACML or WL for a linear mixed effects model (random intercept and slope)
}
