% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{acml}
\alias{acml}
\title{Fit model using ascertainment corrected likelihood model (ACML)}
\usage{
acml(
  formula,
  design,
  data = NULL,
  subset = NULL,
  weights = NULL,
  MI = FALSE,
  MImethod = "direct",
  na.action = getOption("na.action"),
  verbose = 0L,
  init = NULL,
  ...
)
}
\arguments{
\item{formula}{`formula`; (or an object that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of model specification are given under ‘Details’.}

\item{design}{an object of class 'odsdesign'. This specifies the
design of sampling used for the fitting algorithm.}

\item{data}{`data.frame`; an optional data frame, list or environment (or
object coercible by as.data.frame to a data frame) containing the variables
in the model. If not found in data, the variables are taken from
environment(formula), typically the environment from which acml is called.}

\item{subset}{`logical`; an optional vector specifying a subset of
observations to be used in the fitting process. (See additional details
about how this argument interacts with data-dependent bases in the
‘Details’ section of the model.frame documentation.)}

\item{weights}{an optional vector of weights to be used in the fitting
process. Should be NULL or a numeric vector. If non-NULL, weighted least
squares is used with weights weights (that is, minimizing sum(w*e^2));
otherwise ordinary least squares is used. See also ‘Details’,}

\item{MI}{`logical(1)`; Is multiple imputation to be performed. Defaults to FALSE.}

\item{MImethod}{`character(1)`; Specifies multiple imputation method, defaults to 'direct'. Can also be 'indirect'.}

\item{na.action}{`function`; an optional vector specifying a subset of
observations to be used in the fitting process. (See additional details
about how this argument interacts with data-dependent bases in the
‘Details’ section of the model.frame documentation.)}

\item{verbose}{`numeric(1)`; Debugging information printing level
(sent to the optimizer).}

\item{init}{`numeric`; Initial starting position for parameter search
via the optimizer.}

\item{...}{Optional additional parameters passed to sub methods.}
}
\description{
Outcome dependent sampling designs need to be corrected when fitting a
statistical model for proper inferences. This routine will fit and return
the model fit using ascertainment corrected likelihood.
}
\examples{
data(gbti)
design <- ods(Response ~ Month|Patient, 'intercept', p_sample=c(1, 0.25, 1),
              data=gbti, quantiles=c(0.1, 0.9))
est <- acml(Response ~ Month*Genotype, design, gbti)
est
summary(est)
}
