% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds.R
\name{bds}
\alias{bds}
\title{Specify a given design for BLUP Dependent Sampling (BDS)}
\usage{
bds(
  formula,
  method,
  p_sample,
  data = NULL,
  quantiles = NULL,
  cutpoints = NULL,
  subset = NULL,
  na.action = getOption("na.action"),
  ProfileCol = NULL
)
}
\arguments{
\item{formula}{(or an object that can be
coerced to that class): a symbolic description of the sampling
model.}

\item{method}{`character(1)`; A string that specifies 'slope', 'intercept', 'bivariate', or 'mean'.}

\item{p_sample}{`numeric(n)`; A numeric vector with sampling probabilities
for each region specified by cutpoints.}

\item{data}{`data.frame`; an optional data frame, list or environment (or
object coercible by as.data.frame to a data frame) containing the variables
in the model. If not found in data, the variables are taken from
environment(formula), typically the environment from which acml is called.}

\item{quantiles}{`numeric(n-1)`; quantiles to compute cutpoints of interest.
One of quantiles or cutpoints must be specified. Default is NULL.}

\item{cutpoints}{`numeric(n-1)`; a specification of the cutpoints for sampling. Default is NULL.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process. (See additional details about how this
argument interacts with data-dependent bases in the ‘Details’ section of
the model.frame documentation.)}

\item{na.action}{a function which indicates what should happen when the data
contain NAs. The default is set by the na.action setting of options, and is
na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another
possible value is NULL, no action. Value na.exclude can be useful.}

\item{weights}{an optional vector of weights to be used in the fitting
process. Should be NULL or a numeric vector. If non-NULL, weighted least
squares is used with weights weights (that is, minimizing sum(w*e^2));
otherwise ordinary least squares is used. See also ‘Details’,}

\item{...}{additional arguments.}
}
\value{
Returns an BDS design object.
}
\description{
Specify the design of an outcome dependent sampling routine.
}
\examples{
data(gbti)

odsd <- bds(Response ~ Month|Patient, 'mean', p_sample=c(1, 0.25, 1),
            data=gbti, quantiles=c(0.1, 0.9))
summary(odsd)
plot(odsd)

odsd <- bds(Response ~ Month|Patient, 'intercept', p_sample=c(1, 0.25, 1),
            data=gbti, quantiles=c(0.1, 0.9))
summary(odsd)
plot(odsd)

odsd <- bds(Response ~ Month|Patient, 'slope', p_sample=c(1, 0.25, 1),
            data=gbti, quantiles=c(0.1, 0.9))
summary(odsd)
plot(odsd)

odsd <- bds(Response ~ Month|Patient, 'bivariate', p_sample=c(1, 0.25, 1),
            data=gbti, quantiles=c(0.1, 0.9))
summary(odsd)
plot(odsd)

}
\seealso{
[plot.bdsdesign()]
}
