% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{LogLikeC.Score2}
\alias{LogLikeC.Score2}
\title{Calculate the gradient of the conditional likelihood}
\usage{
LogLikeC.Score2(
  y,
  x,
  z,
  w.function,
  id,
  beta,
  sigma.vc,
  rho.vc,
  sigma.e,
  cutpoints,
  SampProb,
  Weights,
  CheeseCalc = FALSE
)
}
\arguments{
\item{y}{response vector}

\item{x}{sum(n_i) by p design matrix for fixed effects}

\item{z}{sum(n_i) by 2 design matric for random effects (intercept and slope)}

\item{id}{sum(n_i) vector of subject ids}

\item{beta}{mean model parameter p-vector}

\item{sigma.vc}{vector of variance components on standard deviation scale}

\item{rho.vc}{vector of correlations among the random effects.  The length should be q choose 2}

\item{sigma.e}{std dev of the measurement error distribution}

\item{cutpoints}{A matrix with the first dimension equal to sum(n_i).  These cutpoints define the sampling regions [bivariate Q_i: each row is a vector of length 4 c(xlow, xhigh, ylow, yhigh); univariate Q_i: each row is a vector of length 2 c(k1,k2) to define the sampling regions, i.e., low, middle, high].  Each subject should have n_i rows of the same values.}

\item{SampProb}{A matrix with the first dimension equal to sum(n_i).   Sampling probabilities from within each region [bivariate Q_i: each row is a vector of length 2 c(central region, outlying region); univariate Q_i: each row is a vector of length 3 with sampling probabilities for each region]. Each subject should have n_i rows of the same values.}

\item{Weights}{Subject specific sampling weights.  A vector of length sum(n_i).  Not used unless using weighted Likelihood}

\item{CheeseCalc}{If FALSE, the function returns the gradient of the conditional log likelihood across all subjects.  If TRUE, the cheese part of the sandwich esitmator is calculated.}
}
\value{
If CheeseCalc=FALSE, gradient of conditional log likelihood.  If CheeseCalc=TRUE, the cheese part of the sandwich estimator is calculated.
}
\description{
Calculate the gradient of the conditional likelihood for the univariate and bivariate sampling cases across all subjects (CheeseCalc=FALSE) or the cheese part of the sandwich estimator if CheeseCalc=TRUE.
}
