% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{LogLikeCAndScore2}
\alias{LogLikeCAndScore2}
\title{Calculate the ascertainment corrected log likelihood and score}
\usage{
LogLikeCAndScore2(params, ProfileCol = NA, Keep.liC = FALSE, subjectData)
}
\arguments{
\item{params}{parameter vector c(beta, log(sigma0), log(sigma1), rho, sigmae)}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maimizing the log likelihood for all other parameters while fixing these columns at the values of params at the location of ProfileCol}

\item{Keep.liC}{If TRUE outputs subject specific conditional log lileihoods to be used for the imputation procedure described in the AOAS paper keep z sum(n_i) by 2 design matric for random effects (intercept and slope)}

\item{subjectData}{Precomputed subject-specific data list from \code{CreateSubjectData}.}
}
\value{
The conditional log likelihood with a "gradient" attribute (if Keep.liC=FALSE) and subject specific contributions to the conditional likelihood if Keep.liC=TRUE).
}
\description{
Calculate the ascertainment corrected log likelihood and score for different designs
}
