% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\name{LogLikeCAndScoreWL}
\alias{LogLikeCAndScoreWL}
\title{Calculate the ascertainment corrected log likelihood and score}
\usage{
LogLikeCAndScoreWL(
  params,
  y,
  x,
  z,
  id,
  w.function,
  cutpoints,
  SampProb,
  Weights,
  ProfileCol = NA,
  Keep.liC = FALSE,
  xcol.phase1,
  ests.phase1
)
}
\arguments{
\item{params}{parameter vector c(beta, log(sigma0), log(sigma1), rho, sigmae)}

\item{y}{response vector}

\item{x}{sum(n_i) by p design matrix for fixed effects}

\item{z}{sum(n_i) by 2 design matric for random effects (intercept and slope)}

\item{id}{sum(n_i) vector of subject ids}

\item{Weights}{Subject specific sampling weights.  A vector of length sum(n_i).  Not used unless using weighted Likelihood}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maimizing the log likelihood for all other parameters while fixing these columns at the values of params at the location of ProfileCol}

\item{Keep.liC}{If TRUE outputs subject specific conditional log lileihoods to be used for the imputation procedure described in the AOAS paper keep z sum(n_i) by 2 design matric for random effects (intercept and slope)}
}
\value{
The conditional log likelihood with a "gradient" attribute (if Keep.liC=FALSE) and subject specific contributions to the conditional likelihood if Keep.liC=TRUE).
}
\description{
Calculate the ascertainment corrected log likelihood and score for different designs
}
