% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODSMethods-package.R
\docType{package}
\name{ODSMethods}
\alias{ODSMethods-package}
\alias{ODSMethods}
\title{Outcome Dependent Sampling (ODS) correction methods}
\description{
When outcome covariates are available one can perform biased
sampling to maximize information returned, or outcome dependent sampling
(ODS). This is especially helpful
to minimize exposure or an expensive test. The statistical analysis
requires correction and this package has routines to assist in this
process via multiple methods.
}
\references{
Jonathan S. Schildcrout, Shawn P. Garbett, and Patrick J. Heagerty. “Outcome Vector Dependent Sampling with Longitudinal Continuous Response Data: Stratified Sampling Based on Summary Statistics”. en. In: Biometrics 69.2 (June 2013), pp. 405–416. issn: 0006-341X, 1541-0420. doi: 10.1111/biom.12013.
 Chiara Di Gravio, Ran Tao, and Jonathan S. Schildcrout. “Design and Analysis of Two-Phase Studies with Multivariate Longitudinal Data”. en. In: Biometrics 79.2 (June 2023), pp. 1420–1432. issn: 0006-341X, 1541-0420. doi: 10.1111/biom.13616.
 Chiara Di Gravio, Jonathan S Schildcrout, and Ran Tao. “Efficient designs and analysis of two-phase studies with longitudinal binary data”. en. In: Biometrics 80.1 (Jan. 2024), ujad010. issn: 0006-341X, 1541-0420. doi: 10.1093/biomtc/ujad010.
 Zhichao Sun et al. “Exposure enriched outcome dependent designs for longitudinal studies of gene–environment interaction”. en. In: Statistics in Medicine 36.18 (Aug. 2017), pp. 2947–2960. issn: 0277-6715, 1097-0258. doi: 10.1002/sim.7332.
 Ran Tao, Donglin Zeng, and Dan-Yu Lin. “Efficient Semiparametric Inference Under Two-Phase Sampling, With Applications to Genetic Association Studies”. en.In: Journal of the American Statistical Association 112.520 (Oct. 2017), pp. 1468–1476. issn: 0162-1459, 1537-274X. doi: 10.1080/01621459.2017.1295864.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vubiostat/ODSMethods}
  \item Report bugs at \url{https://github.com/vubiostat/ODSMethods/issues}
}

}
\author{
\strong{Maintainer}: Lucy Yan \email{bailu.yan@vanderbilt.edu} (\href{https://orcid.org/0000-0002-3718-3117}{ORCID})

Authors:
\itemize{
  \item Shawn Garbett \email{shawn.garbett@vumc.org} (\href{https://orcid.org/0000-0003-4079-5621}{ORCID}) [contributor]
  \item Jonathan Schildcrout \email{jonny.schild@vumc.org} (\href{https://orcid.org/0000-0003-2276-2551}{ORCID}) [contributor]
}

}
\keyword{2-phase,}
\keyword{acml,}
\keyword{ods,}
\keyword{sampling}
