% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODSMethods-package.R
\docType{package}
\name{ODSMethods}
\alias{ODSMethods-package}
\alias{ODSMethods}
\title{Outcome Dependent Sampling (ODS) correction methods}
\description{
When outcome covariates are available one can perform biased
sampling to maximize information returned, or outcome dependent sampling
(ODS). This is especially helpful
to minimize exposure or an expensive test. The statistical analysis
requires correction and this package has routines to assist in this
process via multiple methods.
}
\references{
'Two-Phase, Generalized Case-Control
Designs for the Study of Quantitative Longitudinal Outcomes' Schildcrout,
 et al. (2020) <doi: 10.1093/aje/kwz127>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vubiostat/ODSMethods}
  \item Report bugs at \url{https://github.com/vubiostat/ODSMethods/issues}
}

}
\author{
\strong{Maintainer}: Lucy Yan \email{bailu.yan@vanderbilt.edu} (\href{https://orcid.org/0000-0002-3718-3117}{ORCID})

Authors:
\itemize{
  \item Shawn Garbett \email{shawn.garbett@vumc.org} (\href{https://orcid.org/0000-0003-4079-5621}{ORCID}) [contributor]
  \item Jonathan Schildcrout \email{jonny.schild@vumc.org} (\href{https://orcid.org/0000-0003-2276-2551}{ORCID}) [contributor]
}

}
\keyword{2-phase,}
\keyword{acml,}
\keyword{ods,}
\keyword{sampling}
