% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods.R
\name{plot.odsdesign}
\alias{plot.odsdesign}
\title{Scatter plot an ODS design}
\usage{
\method{plot}{odsdesign}(
  x,
  xlab = "Intercept",
  ylab = "Slope",
  main = format(x$formula),
  sub = paste(x$method, "design"),
  col = if (x$method == "mean") "lightgray" else "black",
  lwd = 1,
  lty = 1,
  cutcol = "red",
  cutlwd = 2,
  cutlty = 2,
  ...
)
}
\arguments{
\item{x}{the ods design object output from \code{\link{ods}}.}

\item{xlab}{a title for the x axis: see \code{\link{title}}.}

\item{ylab}{a title for the y axis: see \code{\link{title}}.}

\item{main}{an overall title for the plot: see \code{\link{title}}.}

\item{sub}{a subtitle for the plot: see \code{\link{title}}.}

\item{col}{A specification for the default plotting color. See \code{\link{par}}.}

\item{lwd}{The line width, a positive number, defaulting to 1. See \code{\link{par}}.}

\item{lty}{The line type. Line types can either be specified as an integer
(0=blank, 1=solid (default), 2=dashed, 3=dotted, 4=dotdash, 5=longdash,
 6=twodash) or as one of the character strings \code{"blank", "solid",
 "dashed", "dotted", "dotdash", "longdash", or "twodash"}, where
 \code{"blank"} uses ‘invisible lines’ (i.e., does not draw them).
 See \code{\link{par}}.}

\item{cutcol}{A specification for the cut point line plotting color.
Defaults to 'red'. See \code{\link{par}}.}

\item{cutlwd}{The cut point line width, a positive number, defaulting to 2.
See \code{\link{par}}.}

\item{cutlty}{The cut point line type. Line types can either be specified as an integer
(0=blank, 1=solid (default), 2=dashed, 3=dotted, 4=dotdash, 5=longdash,
 6=twodash) or as one of the character strings \code{"blank", "solid",
 "dashed", "dotted", "dotdash", "longdash", or "twodash"}, where
 \code{"blank"} uses ‘invisible lines’ (i.e., does not draw them).
 See \code{\link{par}}.}

\item{...}{Additional arguments past to \code{\link{plot}}, \code{\link{hist}},
or \code{\link{lines}} depending on context.}
}
\description{
Generate a basic scatter plot of an ODS design and overlay the
cut points.
}
\seealso{
\code{\link{ods}}
}
