% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acml.R
\name{vi.calc}
\alias{vi.calc}
\title{Calculate V_i = Z_i D t(Z_i) + sig_e^2 I_{n_i}}
\usage{
vi.calc(zi, sigma.vc, rho.vc, sigma.e)
}
\arguments{
\item{zi}{n_i by q design matrix for the random effects}

\item{sigma.vc}{vector of variance components on standard deviation scale}

\item{rho.vc}{vector of correlations among the random effects.  The length should be q choose 2}

\item{sigma.e}{std dev of the measurement error distribution}
}
\value{
V_i
}
\description{
Calculate V_i = Z_i D t(Z_i) + sig_e^2 I_{n_i}
}
