% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WL.R
\name{WL_internal}
\alias{WL_internal}
\title{Fitting function: ACML or WL for a linear mixed effects model (random intercept and slope)}
\usage{
WL_internal(formula, design, InitVals, ProfileCol)
}
\arguments{
\item{InitVals}{starting values for c(beta, log(sigma0), log(sigma1), log((1+rho)/(1-rho)), log(sigmae))}

\item{ProfileCol}{the column number(s) for which we want fixed at the value of param.  Maimizing the log likelihood for all other parameters
while fixing these columns at the values of InitVals[ProfileCol]}

\item{formula.fixed}{formula for the fixed effects (of the form y~x)}

\item{formula.random}{formula for the random effects (of the form ~z).  Right now this model only fits random intercept and slope models.}

\item{data}{data frame that should contain everything in formula.fixed, formula.random, id, and Weights.  It does not include: w.function}

\item{id}{sum(n_i) vector of subject ids (a variable contained in data)}

\item{method}{sum(n_i) vector with possible values that include "mean" (mean of response series), "intercept" (intercept of the regression of Yi ~ zi where zi is the design matrix for the random effects (solve(t.zi* zi) * t.zi)[1,]), "intercept1"  (intercept of the regression of Yi ~ zi where zi is the design matrix for the random effects (solve(t.zi * zi) * t.zi)[1,]). "intercept2" (second intercept of the regression of the Yi ~}

\item{weights}{Subject specific sampling weights.  A vector of length sum(n_i).  This should be a variable in the data dataframe. It should only be used if doing IPWL.  Note if doing IPWL, only use robcov (robust variances) and not covar.  If not doing IPWL, this must be a vectors of 1s.}
}
\value{
Ascertainment corrected Maximum likelihood: Ests, covar, logLik, code, robcov
}
\description{
Fitting function: ACML or WL for a linear mixed effects model (random intercept and slope)
}
